requirements.txt
flask~=3.0.3
flask-cors
cianparser~=1.0.4
scikit-learn~=1.4.2
pandas~=2.2.2
boto3~=1.34.105
python-dotenv~=1.0.1
PyGithub
DateTime~=5.5
joblib~=1.4.2
flask_httpauth